backend/requirements.txt
numpy==1.24.3
tensorflow==2.13.0
opencv-python==4.8.1.78
Pillow==10.2.0
fastapi==0.103.2
uvicorn==0.23.2
python-multipart==0.0.9
typing-extensions==4.5.0
pygame==2.5.2
